+++ booking_api/Models/UserInformation.cs
@@ -7,15 +7,15 @@ namespace BookingApi.Models
     public class UserInformation
     {
         [Key]
-        public int UserInformationId { get; set; }
+        public short UserInformationId { get; set; }  // smallint in database
         
         public string? FirstName { get; set; }
         public string? MiddleName { get; set; }
         public string? LastName { get; set; }
         public string? Email { get; set; }
-        public string? Number { get; set; }
+        public string? Number { get; set; }  // nvarchar(50) in database
         public string? UserCode { get; set; }
-        public short? StatusId { get; set; }
+        public short? StatusId { get; set; }  // smallint in database
         
         // Audit Fields
         public string? CreateUserId { get; set; }
